# openapi/zeus.v1.yaml
openapi: 3.1.0
info:
  title: Zeus API
  description: Enterprise Harness Agent control plane and durable execution API.
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
paths:
  /.well-known/oauth-authorization-server:
    get:
      tags:
      - identity
      operationId: oauth_authorization_server_metadata
      responses:
        '200':
          description: Successful response
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - {}
  /.well-known/openid-configuration:
    get:
      tags:
      - identity
      operationId: oidc_discovery
      responses:
        '200':
          description: Successful response
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - {}
...
  /oauth2/authorize:
    get:
      tags:
      - identity
      operationId: oidc_authorize
      responses:
        '303':
          description: Successful response
        default:
          description: OAuth 2.0 or OpenID Connect error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthProtocolError'
      security:
      - {}
  /oauth2/jwks.json:
    get:
      tags:
      - identity
      operationId: oidc_jwks
      responses:
        '200':
          description: Successful response
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - {}
  /oauth2/logout:
    get:
      tags:
      - identity
      operationId: oidc_logout_get
      responses:
        '303':
          description: Successful response
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - {}
    post:
      tags:
      - identity
      operationId: oidc_logout_post
      responses:
        '303':
          description: Successful response
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - {}
  /oauth2/revoke:
    post:
      tags:
      - identity
      operationId: oidc_revoke
      responses:
        '200':
          description: Successful response
        default:
          description: OAuth 2.0 or OpenID Connect error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthProtocolError'
      security:
      - {}
  /oauth2/token:
    post:
      tags:
      - identity
      operationId: oidc_token
      responses:
        '200':
          description: Successful response
        default:
          description: OAuth 2.0 or OpenID Connect error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthProtocolError'
      security:
      - {}
  /oauth2/userinfo:
    get:
      tags:
      - identity
      operationId: oidc_userinfo_get
      responses:
        '200':
          description: Successful response
        default:
          description: OAuth 2.0 or OpenID Connect error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthProtocolError'
      security:
      - {}
    post:
      tags:
      - identity
      operationId: oidc_userinfo_post
      responses:
        '200':
          description: Successful response
        default:
          description: OAuth 2.0 or OpenID Connect error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthProtocolError'
      security:
      - {}
components:
  schemas:
...
    OAuthProtocolError:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        error_description:
          type:
          - string
          - 'null'
...
    ProblemDetails:
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      - request_id
      properties:
        code:
          type: string
        detail:
          type: string
        request_id:
          type: string
          format: uuid
        status:
          type: integer
          format: int32
          minimum: 0
        title:
          type: string
        type:
          type: string
          example: https://zeus.example.com/problems/validation_failed
...
  securitySchemes:
    serviceAccountBearer:
      type: http
      scheme: bearer
      bearerFormat: ServiceAccount
      description: Bearer token for a Zeus service account
    sessionCookie:
      type: apiKey
      in: cookie
      name: zeus_session
      description: HttpOnly session cookie
...